Split Generic Fixtures
