+++ detekt/baseline.xml
@@ -12,10 +12,9 @@
     <ID>MagicNumber:build.gradle.kts$8</ID>
     <ID>MagicNumber:settings.gradle.kts$30</ID>
     <ID>MaxLineLength:Dangerfile.df.kts$val</ID>
-    <ID>MaxLineLength:Fixture.kt$public</ID>
     <ID>MaxLineLength:GradlePlugin.kt$GradlePlugin$const val runtimeConfig = "tech.antibytes.gradle-plugins:antibytes-runtime-configuration:${Version.Gradle.antibytes}"</ID>
+    <ID>MaxLineLength:TupleFixture.kt$public</ID>
     <ID>MaxLineLength:build.gradle.kts$config = files("$projectDir/detekt/config.yml")</ID>
-    <ID>TooManyFunctions:Fixture.kt$tech.antibytes.kfixture.Fixture.kt</ID>
     <ID>TooManyFunctions:RandomWrapper.kt$RandomWrapper : Random</ID>
     <ID>UnusedPrivateMember:build.gradle.kts$val androidMain by getting { dependencies { dependsOn(commonMain) implementation(Dependency.multiplatform.kotlin.android) } }</ID>
     <ID>UnusedPrivateMember:build.gradle.kts$val androidTest by getting { dependsOn(androidAndroidTestRelease) dependsOn(androidTestFixtures) dependsOn(androidTestFixturesDebug) dependsOn(androidTestFixturesRelease) dependencies { implementation(Dependency.multiplatform.test.jvm) implementation(Dependency.multiplatform.test.junit) } }</ID>
